+++ src/main/java/spring/jpa/board/controller/UserRestController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import spring.jpa.board.dto.ApiResponse;
 import spring.jpa.board.dto.user.UserDto;
 import spring.jpa.board.service.UserService;
