+++ src/main/java/spring/jpa/board/dto/ApiResponse.java
@@ -1,4 +1,4 @@
-package spring.jpa.board.controller;
+package spring.jpa.board.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
